+++ tests/test_mydevolo.py
@@ -1,6 +1,6 @@
 import pytest
 
-from devolo_home_control_api.mydevolo import Mydevolo, WrongUrlError, WrongCredentialsError
+from devolo_home_control_api.mydevolo import Mydevolo, GatewayOfflineError, WrongUrlError, WrongCredentialsError
 
 
 class TestMydevolo:
@@ -108,6 +108,18 @@ class TestMydevolo:
         with pytest.raises(WrongUrlError):
             mydevolo._call("test")
         Mydevolo.del_instance()
+
+    @pytest.mark.usefixtures("mock_mydevolo__call_raise_GatewayOfflineError")
+    def test_call_GatewayOfflineError(self):
+        mydevolo = Mydevolo()
+        with pytest.raises(GatewayOfflineError):
+            mydevolo._call("test")
+        Mydevolo.del_instance()
+
+    @pytest.mark.usefixtures("mock_response_gateway_offline")
+    def test_call_GatewayOfflineError(self, mydevolo):
+        with pytest.raises(GatewayOfflineError):
+            mydevolo._call("test")
 
     @pytest.mark.usefixtures("mock_response_valid")
     def test_call_valid(self, mydevolo):